+++ noodles-fastq/src/async.rs
@@ -1,6 +1,3 @@
 //! Async FASTQ.
 
 pub mod io;
-mod writer;
-
-pub use self::writer::Writer;