+++ noodles-fastq/src/async/io.rs
@@ -1,5 +1,6 @@
 //! Async FASTQ I/O.
 
 mod reader;
+mod writer;
 
-pub use self::reader::Reader;
+pub use self::{reader::Reader, writer::Writer};