+++ noodles-fastq/src/io.rs
@@ -1,5 +1,6 @@
 //! FASTQ I/O.
 
 pub mod reader;
+mod writer;
 
-pub use self::reader::Reader;
+pub use self::{reader::Reader, writer::Writer};