+++ noodles-fastq/CHANGELOG.md
@@ -4,13 +4,13 @@
 
 ### Changed
 
-  * fastq: Move reader (`Reader`) to `io` module.
+  * fastq: Move reader (`Reader`) and writer (`Writer`) to `io` module.
 
 ### Deprecated
 
-  * fastq: Deprecate `fastq::Reader`.
+  * fastq: Deprecate `fastq::Reader` and `fastq::Writer`.
 
-    This is moved under the `io` module.
+    These are moved under the `io` module.
 
 ## 0.10.0 - 2023-12-14